LPS-85893 SF

+++ portal-impl/src/com/liferay/portal/repository/liferayrepository/model/LiferayFileShortcut.java
@@ -255,7 +255,7 @@ public class LiferayFileShortcut extends LiferayModel implements FileShortcut {
 
 	@Override
 	public void setPrimaryKeyObj(Serializable primaryKeyObj) {
-		setPrimaryKey(((Long)primaryKeyObj).longValue());
+		setPrimaryKey((Long)primaryKeyObj);
 	}
 
 	@Override
