+++ portal-impl/src/com/liferay/portal/repository/liferayrepository/model/LiferayFolder.java
@@ -383,7 +383,7 @@ public class LiferayFolder extends LiferayModel implements Folder {
 
 	@Override
 	public void setPrimaryKeyObj(Serializable primaryKeyObj) {
-		setPrimaryKey(((Long)primaryKeyObj).longValue());
+		setPrimaryKey((Long)primaryKeyObj);
 	}
 
 	@Override
